Add parsing of the periodic timestamps to the transcript API
I need to be able to search transcripts to lookup the time at which
various events happened.

Campfire transcripts don't have exact timestamps.  But there is a
timestamp inserted every five minutes or so, which can be used to give
a rough sense of time.  Currently, the Tinder API doesn't provide
any way to access this timing information.

Make that approximate timestamp available as part of each message hash
in a transcript until the next timestamp is encountered.