Fix scraping of the person who entered or left the room in transcripts
I need to know who entered and left a room in a transcript.

Currently, Campfire omits the span tag around the person's name when
rendering the enter/leave messages in the transcript.  As a result,
the logic for parsing regular messages doesn't work.

So just use the top-level ".person" element if ".person span" is
empty.